+++ src/Configuration.java
@@ -1,48 +1,48 @@
 public class Configuration {
-    private int totalPages;      // p: número de páginas do espaço de endereçamento virtual
-    private int totalFrames;     // m: número de molduras da memória
-    private int clockInterval;   // c: ciclo de relógio do bit R
-    private int pageNumber;      // número da página sendo acessada
-    private int accessTime;      // t: momento do acesso
-    private boolean isWrite;     // true para W (escrita), false para R (leitura)
+    private long totalPages;      
+    private long totalFrames;     
+    private long clockInterval;   
+    private long pageNumber;      
+    private long accessTime;      
+    private boolean isWrite;     
 
-    public int getTotalPages() {
+    public long getTotalPages() {
         return totalPages;
     }
 
-    public void setTotalPages(int totalPages) {
+    public void setTotalPages(long totalPages) {
         this.totalPages = totalPages;
     }
 
-    public int getTotalFrames() {
+    public long getTotalFrames() {
         return totalFrames;
     }
 
-    public void setTotalFrames(int totalFrames) {
+    public void setTotalFrames(long totalFrames) {
         this.totalFrames = totalFrames;
     }
 
-    public int getClockInterval() {
+    public long getClockInterval() {
         return clockInterval;
     }
 
-    public void setClockInterval(int clockInterval) {
+    public void setClockInterval(long clockInterval) {
         this.clockInterval = clockInterval;
     }
 
-    public int getPageNumber() {
+    public long getPageNumber() {
         return pageNumber;
     }
 
-    public void setPageNumber(int pageNumber) {
+    public void setPageNumber(long pageNumber) {
         this.pageNumber = pageNumber;
     }
 
-    public int getAccessTime() {
+    public long getAccessTime() {
         return accessTime;
     }
 
-    public void setAccessTime(int accessTime) {
+    public void setAccessTime(long accessTime) {
         this.accessTime = accessTime;
     }
